backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pydantic==2.11.2
google-genai==1.14.0
python-dotenv==1.1.0
aiofiles==24.1.0
